Move compute_turning_angle from orientation to path
