+++ movement/kinematics/orientation.py
@@ -7,10 +7,8 @@ import numpy as np
 import xarray as xr
 from numpy.typing import ArrayLike
 
-from movement.kinematics.kinematics import compute_backward_displacement
 from movement.utils.logging import logger
 from movement.utils.vector import (
-    compute_norm,
     compute_signed_angle_2d,
     convert_to_unit,
 )
@@ -281,121 +279,3 @@ def _validate_type_data_array(data: xr.DataArray) -> None:
                 f"but got {type(data)}."
             )
         )
-
-
-def compute_turning_angle(
-    data: xr.DataArray,
-    in_degrees: bool = False,
-    min_step_length: float = 0.0,
-) -> xr.DataArray:
-    r"""Compute the turning angles between consecutive steps in a trajectory.
-
-    The turning angle at time ``t`` is  the :func:`signed angle\
-    <movement.utils.vector.compute_signed_angle_2d>` between two
-    consecutive :func:`backward displacement\
-    <movement.kinematics.compute_backward_displacement>` vectors
-    at times ``t-1`` and ``t``.
-    The returned angles are in radians, spanning the range :math:`(-\pi, \pi]`,
-    unless ``in_degrees`` is set to ``True``.
-
-    Parameters
-    ----------
-    data
-        The input position data. Must contain ``time`` and ``space``
-        dimensions. The ``space`` dimension must contain exactly the
-        coordinates ``["x", "y"]`` (2D spatial data only).
-    in_degrees
-        If ``True``, return turning angles in degrees. Default is
-        ``False`` (radians).
-    min_step_length
-        The minimum step length to consider for computing the turning
-        angle. Any turning angle involving an incoming or outgoing step
-        shorter than or equal to this value is set to ``NaN``. The
-        default ``0.0`` only masks steps with exactly zero length,
-        which means steps with near-zero lengths may still produce
-        spurious angles. See Note 2 below.
-
-    Returns
-    -------
-    xr.DataArray
-        Turning angles with the same shape as the input ``data``, but
-        with the ``space`` dimension dropped.
-
-    Notes
-    -----
-    1. **Time dimension length:** This function uses a ``shift``
-       operation to preserve the original ``time`` dimension length.
-       The first two time steps are always ``NaN``: the first because
-       no previous step exists, and the second because a turning angle
-       requires two steps (three positions). In other words, the turning angle
-       at time step ``t`` is computed as the angle between the steps
-       from ``t-2`` to ``t-1`` and from ``t-1`` to ``t``.
-    2. **Positional jitter and small steps:** Tracking data
-       often contains positional jitter, meaning a stationary animal
-       may appear to make microscopic movements. With default parameters
-       (``min_step_length=0.0``), these tiny, noisy movements will
-       produce spurious, meaningless turning angles. It is highly
-       recommended to set ``min_step_length`` to an appropriate threshold
-       based on the tracking resolution and the animal's size in the scene.
-       The value should be in the same units as the input position data
-       (e.g. pixels, mm, etc.). Pre-smoothing the trajectory
-       can also help reduce positional jitter.
-    3. **NaN propagation:** ``NaN`` positions in the input propagate
-       to ``NaN`` turning angles. A single missing position affects
-       up to two turning angles (the incoming and outgoing steps).
-       Use :func:`movement.filtering.interpolate_over_time` to fill
-       positional gaps before computing turning angles if continuity
-       is important.
-
-    See Also
-    --------
-    movement.kinematics.compute_backward_displacement :
-        The underlying function used to compute the displacement vectors.
-    movement.utils.vector.compute_signed_angle_2d :
-        The underlying function used to compute the signed angle
-        between two consecutive displacement vectors.
-
-    Examples
-    --------
-    >>> from movement.kinematics import compute_turning_angle
-
-    Compute turning angles from the centroid trajectory of a poses
-    dataset ``ds``:
-
-    >>> centroid = ds.position.mean(dim="keypoint")
-    >>> angles = compute_turning_angle(centroid)
-
-    Compute in degrees, with a minimum step length of 3 pixels to filter out
-    pose estimation jitter:
-
-    >>> angles = compute_turning_angle(
-    ...     centroid, in_degrees=True, min_step_length=3
-    ... )
-
-    """
-    validate_dims_coords(
-        data, {"time": [], "space": ["x", "y"]}, exact_coords=True
-    )
-
-    # Displacement arriving at each time step t.
-    disp = compute_backward_displacement(data)
-
-    # Turning angle at t = rotation needed to align step[t-1] onto step[t].
-    turning = compute_signed_angle_2d(disp.shift(time=1), disp)
-
-    # Mask turning angles involving steps smaller than min_step_length
-    step_lengths = compute_norm(disp)
-    invalid_steps = (step_lengths <= min_step_length) | (
-        step_lengths.shift(time=1) <= min_step_length
-    )
-    turning = xr.where(invalid_steps, np.nan, turning)
-
-    turning.attrs["units"] = "radians"
-
-    if in_degrees:
-        turning = np.rad2deg(turning)
-        turning.attrs["units"] = "degrees"
-
-    turning.name = "turning_angle"
-
-    return turning
